feat: Import to database extracted from StatisticsApplication to DatabaseImporter. Excluded JPA configuration from StatisticsApplication. - All components are set to "database" profile. - We can now execute PlayerDataDownloader via StatisticsApplication, but JPA configuration will fail, i.e. DB import and generation will fail.
