+++ .run/StatisticsApplication - IMPORT_JSON_TO_DATABASE.run.xml
@@ -3,6 +3,7 @@
     <option name="MAIN_CLASS_NAME" value="ru.klavogonki.statistics.springboot.StatisticsApplication" />
     <module name="kgstats-srv" />
     <option name="PROGRAM_PARAMETERS" value="IMPORT_JSON_TO_DATABASE C:\java\kgparser\kgstatsSrv\src\main\resources\statistics\config-output.json" />
+    <option name="VM_PARAMETERS" value="-Dspring.profiles.active=database" />
     <extension name="coverage">
       <pattern>
         <option name="PATTERN" value="ru.klavogonki.statistics.export.vocabulary.standard.*" />
