+++ kgstatsSrv/src/main/java/ru/klavogonki/statistics/export/PlayersByRankExporter.java
@@ -2,12 +2,14 @@ package ru.klavogonki.statistics.export;
 
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
+import ru.klavogonki.statistics.dto.PlayerRankLevelAndTotalRacesCount;
 import ru.klavogonki.statistics.freemarker.PageUrls;
 import ru.klavogonki.statistics.freemarker.PlayersByRankDataTemplate;
 import ru.klavogonki.statistics.freemarker.PlayersByRankTemplate;
-import ru.klavogonki.statistics.dto.PlayerRankLevelAndTotalRacesCount;
 import ru.klavogonki.statistics.repository.PlayerRepository;
+import ru.klavogonki.statistics.springboot.Profiles;
 import ru.klavogonki.statistics.util.JacksonUtils;
 
 import java.util.Comparator;
@@ -15,6 +17,7 @@ import java.util.List;
 
 @Log4j2
 @Component
+@Profile(Profiles.DATABASE)
 public class PlayersByRankExporter implements DataExporter {
 
     private static final int MIN_TOTAL_RACES_COUNT = 1;
