+++ kgstatsSrv/src/main/java/ru/klavogonki/statistics/export/vocabulary/non_standard/OneHundredRussianTopExporter.java
@@ -2,12 +2,15 @@ package ru.klavogonki.statistics.export.vocabulary.non_standard;
 
 import lombok.extern.log4j.Log4j2;
 import org.apache.logging.log4j.Logger;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 import ru.klavogonki.kgparser.NonStandardDictionary;
 import ru.klavogonki.statistics.export.vocabulary.standard.NonStandardVocabularyTopExporterDefaultImpl;
+import ru.klavogonki.statistics.springboot.Profiles;
 
 @Log4j2
 @Component
+@Profile(Profiles.DATABASE)
 public class OneHundredRussianTopExporter extends NonStandardVocabularyTopExporterDefaultImpl {
 
     @Override
