+++ kgstatsSrv/src/main/java/ru/klavogonki/statistics/export/vocabulary/standard/NoErrorTopExporter.java
@@ -2,11 +2,14 @@ package ru.klavogonki.statistics.export.vocabulary.standard;
 
 import lombok.extern.log4j.Log4j2;
 import org.apache.logging.log4j.Logger;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 import ru.klavogonki.kgparser.StandardDictionary;
+import ru.klavogonki.statistics.springboot.Profiles;
 
 @Log4j2
 @Component
+@Profile(Profiles.DATABASE)
 public class NoErrorTopExporter extends StandardVocabularyTopExporterDefaultImpl {
 
     @Override
