+++ kgstatsSrv/src/main/java/ru/klavogonki/statistics/export/vocabulary/standard/NonStandardVocabularyTopExporterDefaultImpl.java
@@ -1,17 +1,19 @@
 package ru.klavogonki.statistics.export.vocabulary.standard;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
-import ru.klavogonki.statistics.export.vocabulary.non_standard.NonStandardVocabularyTopExporter;
 import ru.klavogonki.statistics.entity.PlayerEntity;
 import ru.klavogonki.statistics.entity.PlayerVocabularyStatsEntity;
+import ru.klavogonki.statistics.export.vocabulary.non_standard.NonStandardVocabularyTopExporter;
 import ru.klavogonki.statistics.repository.PlayerVocabularyStatsRepository;
+import ru.klavogonki.statistics.springboot.Profiles;
 
 import java.util.List;
 
 // todo: same implementation as StandardVocabularyTopExporterDefaultImpl
 @Component
+@Profile(Profiles.DATABASE)
 public abstract class NonStandardVocabularyTopExporterDefaultImpl implements NonStandardVocabularyTopExporter {
 
     @Autowired
