+++ kgstatsSrv/src/main/java/ru/klavogonki/statistics/springboot/Profiles.java
@@ -1,0 +1,9 @@
+package ru.klavogonki.statistics.springboot;
+
+public final class Profiles {
+
+    private Profiles() {
+    }
+
+    public static final String DATABASE = "database";
+}
